[rom_ctrl] Split ROM index to catch FI on address lines
Before this change, there were some unprotected address line signals
between the TL -> SRAM adapter and the ROM. If an attacker managed to
inject a fault into one of these lines, they could redirect a fetch
and potentially redirect the ROM code. This would probably be bad!

This commit splits the ROM addresses that we pass around into two so
that the scrambled ROM now takes two addresses. It uses one to index
into the physical ROM and the other to derive the (address-tweakable)
keystream. The idea is that if you inject a fault on only one of these
addresses then you'll get garbage because it will descramble to the
wrong value.

Note that we don't bother splitting the addresses that come out of the
checker FSM: there are probably easier ways to redirect that anyway
and they won't help an attacker because the result will fail the ROM
integrity check.

For bus accesses, we split out one of the addresses "upstream" of the
TL -> SRAM adapter. Because that adapter contains an ECC check, we'd
expect to detect any FI attack that affected both sides of the fork.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>