[spi_device] Fix assertion condition TpmEnable
The intention of `TpmEnableWhenTpmCsbIdle_M` assumption is to check the
SW sets tpm_en only when the traffic is in idle. The qualifier condition
should be `$rose(cfg_tpm_en)` to catch the write moment.

In previous design, it used `cfg_tpm_en` directly, so after the bit is
set, assumption fired at every cycle.

The issue is reported by @kosta-kojdic in #9629

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>