[reggen] Tweak Register's _asdict to satisfy new mypy
Mypy 0.920 was released last night and now infers a more precise type
for the "rd" object that we return. Instead of Dict[str, object], it
infers Dict[str, Sequence[object]] (presumably because we happen to
always use a string or a list as a value). Because Dict is not
covariant in the value type, mypy then complains that we're trying to
return rd from the function.

The fix is to explicitly tell mypy what type to infer for rd.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>