[ entropy_src, rtl ] Correct bucket health test test_cnt_o
The previous version of entropy_src_bucket_ht output a count of failing tests on test_cnt_o
This is not what is specified in the documentation. Rather this is supposed to output the
value of the most heavily weighted bin.

This commit calculates the maximum of the (2 ** RngBusWidth) histogram bins, by creating
a tree of comparators (RngBusWidth deep).

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>