[aes, pre_syn] Remove prim_sparse_fsm_flop suffixes introduced by sv2v
It seems that related to the StateEnumT parameter of the
prim_sparse_fsm_flop module, sv2v introduces some additional suffixes to
to the module and instantiations that break Yosys synthesis. Based on
the sv2v version used, these suffixes might vary. This commit removes
additional variations that we've seen so far.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>