[sw,crypto] Adjust parameters for R^2 algorithm.
Because the operands in this computation are always powers of 2, it is
equivalent to square the operand or to perform a certain number of
(modular) doublings. For small numbers the doublings are faster and for
larger ones the squaring is faster; the cutoff where this happens for
our implementation is at 2^96 (empirically determined).

Signed-off-by: Jade Philipoom <jadep@google.com>