[dif/alert_handler] Add alert configuration DIF.
Previously, the only way to configure an alert within the alert handler
was to build a rather large struct that would configure all alerts
within a class. (Re)Configuring a single alert had the side of effect of
reconfiguring the class too.

This commit adds a DIF to configure a single alert, i.e., enable it, set
its class, and lock it. This partially addresses #9899.

Signed-off-by: Timothy Trippel <ttrippel@google.com>