[aes/syn] Update GTECH flow for AES wrapper
This separate synthesis target for the AES wrapper is only intended to
produce a mapping of the wrapper alone. The motivation for synthesizing
AES and the wrapper separately is that we can ensure that no sort of
optimization takes place across the hierarchy boundary between them (the
wrapper is TB code whereas the AES is actual synthesizable code).

In order to achieve this we just remove the aes module from the working
lib after elaboration.

Signed-off-by: Michael Schaffner <msf@google.com>