[otbn] Fix qualifying writes to registers by operational state
We could write to CMD even when in the locked state (oops!). I've also
tidied things up a little, defining an 'idle' signal so that the
current operational state is one-hot encoded in

    {busy_execute_q, locked, idle}

I've also defined an 'is_not_running' signal that we use to for
idle_o ("clock gate me") and to qualify writes to ERR_BITS, INSN_CNT.

Amusingly, the DV code had the same problem so we have to do
corresponding updates for the scoreboard and a checker for the
"idle_o" output as well.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>