[flash_ctrl] Allow rma entry even if flash has not been initialized.
As part of #10188, it was realized that rma entry is currently gated
by flash initilization.

This may not be a good idea because if flash initlization were to ever
stall (ROM could not program it for whatever reason), then rma entry
could never be done.

This PR allows RMA entry to happen even if initilization has not been
called.  If initilization has not been called, it is possible that
entropy is not properly behaving as part of the ROM routine, so the
reseeding is also skipped so that the flash can be directly wiped.

Signed-off-by: Timothy Chen <timothytim@google.com>