[usbdev] Move sense to an MIO
The ASIC was missing a connection for the VBUS detection, so the MIO
adds that. In addition, the FPGAs connect the two VBUS detection sources
to IOR0 and IOR1--Software will have to select the correct I/O based on
the `use_uphy` value.

Moving to an MIO brings straightforward options for a simple constant
via periph_insel if OT is completely bus-powered, in addition to
making it available to both power domains for usbdev.

Signed-off-by: Alexander Williams <awill@google.com>